requirement.txt
customtkinter>=5.2.2
yara-python>=4.3.1